azure_functions/requirements.txt
# Azure Functions runtime package
azure-functions

# MySQL connector for Python
mysql-connector-python

# Optional (recommended for cloud logging / structured logs)
azure-core
